Add files via upload
This APMC (Agriculture Produce Market Committee) project has 3 Login namely CLERK, USER and ADMIN.  The Admin is responsible to add menu, edit menu, display specific menu according to the logged user. The CLERK is responsible to manage all the transactions within the APMC. AVAK, CESS, KATAPATTI, SAUDAPATTI , JAVAK these all are the attached webforms which are managed by the CLERK of specific APMC. The User can perform the operations such as Registering on the Website to get the license of the APMC. The user can be Farmer, Purchaser, etc.